refactor : 삭제될 때 history에 넣음

+++ src/main/java/qna/domain/Answers.java
@@ -17,9 +17,10 @@ public class Answers {
     public List<DeleteHistory> delete(User loginUser) throws CannotDeleteException {
         List<DeleteHistory> deleteHistories = new ArrayList<>();
         for (Answer answer : answers) {
-            answer.deleteByOwner(loginUser);
-            DeleteHistory deleteHistory = new DeleteHistory(ContentType.ANSWER, answer.getId(), loginUser, LocalDateTime.now());
-            deleteHistories.add(deleteHistory);
+            if(answer.deleteByOwner(loginUser)){
+                DeleteHistory deleteHistory = new DeleteHistory(ContentType.ANSWER, answer.getId(), loginUser, LocalDateTime.now());
+                deleteHistories.add(deleteHistory);
+            }
         }
         return deleteHistories;
     }
